feat: replace management chat with kanban board
